Move the handshake and http modules out of the way.
