+++ docs/changelog.rst
@@ -16,6 +16,10 @@ Changelog
       :exc:`~datastructures.MultipleValuesError` from
       :mod:`websockets.datastructures` instead of :mod:`websockets.http`.
 
+    * :mod:`websockets.handshake` is deprecated.
+
+    * :mod:`websockets.http` is deprecated.
+
     Aliases provide backwards compatibility for all previously public APIs.
 
 8.1
